Minor doc improvement to navigation section in README.md
In its current place (after outlining the form route definitions), `set_url_hash('home')` was implying that the `home` would be a path available after the `HomeForm` was decorated. But `home` is an explicit page name that needs to be defined.

In this commit, I have replaced the `home` page name with a more generic `articles` string and also added a separate example for `home`.